backend/requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
groq==0.36.0
pydantic==2.12.5
python-dotenv==1.2.1
sentence-transformers==5.1.2
numpy==2.3.5
scikit-learn==1.7.2
torch==2.9.1
